feat(agent): fleet-agent が Quadlet backend の deploy に対応 (WS3)
Podman+Quadlet 追従 epic (mem_1CbD3b6j1s3pxQ1TGvaXtv) WS3。

WS2 Stage 2c で `fleet up` をホスト上で実行した時の Quadlet 経路を作った。
WS3 はそれを fleet-agent にやらせる — CP→agent 経由のクラウドデプロイで
agent が Quadlet を適用する。

## 共有 crate への集約

`build_stage_units` / `service_units` を CLI crate (commands/quadlet.rs)
から `fleetflow-container::quadlet` に移動（agent も使うため）。さらに
「build → sync → daemon-reload → start」を `apply_stage` に集約:
- `apply_stage(flow, stage_name, stage) -> QuadletApplyOutcome`
- CLI・agent の両方が共用、各々は薄い presenter

## agent 側

`handle_deploy_kdl` が stage の backend で分岐:
- `Backend::Quadlet` → `apply_stage`（bollard 不使用）
- `Backend::Compose` → 未実装エラー（WS4）
- `Backend::Docker` → 従来の DeployEngine + bollard 経路

CP↔agent のコマンド応答プロトコル（WS0 で修正済）と DeployRequest
（Flow + stage）はそのまま再利用。

## commands/quadlet.rs

inline ロジックを `apply_stage` 呼び出しに置換、薄い presenter 化。

注: `systemctl --user` は呼び出しプロセスのユーザーセッションに作用する。
agent が rootless Quadlet を適用するには agent プロセスが対象 rootless
ユーザーである必要がある（provisioning は WS5）。

テスト: build_stage_units 系 4 件を fleetflow-container に移設。
